Auto-skip systemd units

+++ src/core.nix
@@ -56,29 +56,38 @@ let
       pkgs.coreutils
       pkgs.sd-switch
     ];
-    text = ''
-      config_home="''${XDG_CONFIG_HOME:-$HOME/.config}"
-      echo ":: Loading systemd-tmpfiles"
-      mkdir -p "$config_home/systemd"
-      mkdir -p "$config_home/user-tmpfiles.d"
-      ln -sfT "${config.build.staticTmpfiles}/lib/nix-maid/00-nix-maid-tmpfiles.conf" "$config_home/user-tmpfiles.d/00-nix-maid-tmpfiles.conf"
-      '${lib.getExe config.build.tmpfileRenderer}' > "$config_home/user-tmpfiles.d/00-nix-maid-dynamic-tmpfiles.conf"
-      '${config.systemd.package}/bin/systemd-tmpfiles' --user --create --remove
+    text =
+      ''
+        config_home="''${XDG_CONFIG_HOME:-$HOME/.config}"
+        echo ":: Loading systemd-tmpfiles"
+        mkdir -p "$config_home/systemd"
+        mkdir -p "$config_home/user-tmpfiles.d"
+        ln -sfT "${config.build.staticTmpfiles}/lib/nix-maid/00-nix-maid-tmpfiles.conf" "$config_home/user-tmpfiles.d/00-nix-maid-tmpfiles.conf"
+        '${lib.getExe config.build.tmpfileRenderer}' > "$config_home/user-tmpfiles.d/00-nix-maid-dynamic-tmpfiles.conf"
+        '${config.systemd.package}/bin/systemd-tmpfiles' --user --create --remove
+      ''
+      + (
+        if (config.systemd.units == { }) then
+          ''
+            echo ":: No systemd.<name> option set, skipping systemd units"
+          ''
+        else
+          ''
+            # Init empty array
+            sd_switch_flags=()
+            # Check if it's a symlink
+            if [[ -h "$config_home/systemd/user" ]]; then
+              sd_switch_flags+=("--old-units" "$(realpath "$config_home/systemd/user")")
+            elif [[ -e "$config_home/systemd/user" ]]; then
+              rm -rf "$config_home/systemd/user"
+            fi
 
-      # Init empty array
-      sd_switch_flags=()
-      # Check if it's a symlink
-      if [[ -h "$config_home/systemd/user" ]]; then
-        sd_switch_flags+=("--old-units" "$(realpath "$config_home/systemd/user")")
-      elif [[ -e "$config_home/systemd/user" ]]; then
-        rm -rf "$config_home/systemd/user"
-      fi
+            ln -sfT "${config.build.units}" "$config_home/systemd/user"
 
-      ln -sfT "${config.build.units}" "$config_home/systemd/user"
-
-      echo ":: Loading systemd units"
-      sd-switch --new-units "$config_home/systemd/user" "''${sd_switch_flags[@]}"
-    '';
+            echo ":: Loading systemd units"
+            sd-switch --new-units "$config_home/systemd/user" "''${sd_switch_flags[@]}"
+          ''
+      );
   };
 in
 {
